tests/wpt/tests/svg/path/parsing/path-length-invalid.tentative.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml">
  <title>SVG Path Properties: parsing path-length with invalid values</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/paths.html#PathLengthProperty"/>
    <h:meta name="assert" content="path-length supports only the grammar 'none | &lt;number [0,∞]&gt;'."/>
  </metadata>
  <path id="target"/>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_invalid_value("path-length", "-1");
test_invalid_value("path-length", "-0.5");
test_invalid_value("path-length", "auto");
test_invalid_value("path-length", "10px");
test_invalid_value("path-length", "10%");
test_invalid_value("path-length", "1 2");
test_invalid_value("path-length", "none 1");

  ]]></script>
</svg>
